Add generic way for defining custom exporters in configuration
By adding `ArticleJSON::Configuration#register_element_exporters_for`
we can in the future allow the user to use the same method to also add
custom exporters e.g. for AMP.

Adapt and deprecate the existing methods so that they still work as they
currently do but can be removed in the future.