Add support for placing additional elements into article
The new method `Article#place_additional_elements` allows a user to
place additional elements into the article, like e.g. advertisements.
These elements need to be defined within the user's own code.

Each element that is added this way will directly get placed in between
paragraphs of the article. The method ensures that an additional element
is never added before or after any node other than paragraphs (e.g. an
image). The elements are added in the order you pass them into the
method. If the article should not have enough spaces to place all the
provided elements, they will be placed after the last element in the
article.

Usage examples are provided in the updated `README.md`.