Allow passing a block to `create_node`
This avoids having to use `.tap` everywhere in the code. It also allows
in future to wrap the output of `create_element` into `NodeSet`s.

Also fix the annotations of the methods: The return value is always just
`Nokogiri::XML::Element`,  `Nokogiri::HTML::Node` does not even exist
within Nokogiri.