Render message for bad OEmbed responses
When a requested OEmbed object returns an error, render a message for
the user explaining that there was an error with the embedded element.

For this, standardize the `OEbedResolver::Base#source_url` method and
add a `#name` method which will be included in the text for the user.

The text for the user will be rendered within the `<figure>` element
within a nested `span` with the class `unavailable-embed`. This also
allows the consumer to have custom CSS for this case.