Export Facebook text boxes surrounded by a horizontal line
Facebook does not support any text box element nor does it allow having
a custom CSS for styling a given div / class (like our `text-box`
class). Text boxes normally contain content that is not part of the main
article, but has a supportive character. Therefore, it is important that
the reader understands that the main content continues after this
paragraph.

This commit inserts a horizontal line above and below the actual text
box content which should provide this visual clue of "to-be-continued".

Floating content is not supported, therefore don't add any floating
classes.