Add rubocop and codeclimate config files
We are running rubocop on codeclimate and until now have used the
default configuration. But the default configuration for the rubocop
inspections on codeclimate are checking for Ruby 2.1 syntax. Since this
gem makes use of new Ruby 2.3 syntax elements (e.g. the safe navigation
operator `&.`), we need to manually configure this.

The configuration was downloaded from codeclimate.com (link below) and
then adapted for our needs.

Source:
https://codeclimate.com/snapshots/5a0437c621d8e4000102a1e9/inferred_config