Add a plain text exporter
As the name suggests, this exporter generates a plain text version of
the article. Rich text elements like images, embeds or even text boxes
are not being rendered.

For this, extract the reusable part of the `Base` concern and implement
the element exporters for this format. There are no exporters for the
skipped items. But via the custom exporter API it is easy to add them,
if they should be required for some consumers.