Improve multiple calls to `Article#place_additional_elements`
When calling `ArticleJSON::Article#place_additional_elements` multiple
times with additional elements, they were previously directly placed
into the article. This led to additional elements being placed very
close together, since the second iteration did not know that there was
already an additional element inserted earlier.

Now, the class keeps track of what additional elements are to be placed
into the article, but only when accessing the `Article#elements` they
are actually injected into the article. If more additional elements are
added, the structure will be recalculated to guarantee an even
distribution of the additional elements.