Fix additional element placement on empty articles
If an article contains no elements but additional elements are place on
it, it previously introduced a `nil` element into the article elements.
This in turn then led the exporters to crash.

Now we check if there are any elements in the original article and if
not, the element place simply returns the additional elements.