Import image href from caption text using a custom tag
Although google docs lets the user to create images with an href, it doesn't
export them to html.
(See https://productforums.google.com/forum/#!topic/docs/85jCFF6McsU)

As a workaround for this, we create a custom tag [image-link-to: url] that when
used in the next paragraph of an image (which we actually use as the caption of
the image) we extract the url and pass it to the Image Element.

Next step would be to export this to the different formats.