Remove facebook videos from reference documents
Facebook changed it's oembed api and now it requires an access token, to get it
you need a facebook account and a facebook app.

Although we may mantain facebook videos, we are not using them in the reference
documents, so that we can tests all the different export formats without a
facebook account.