Fix URL for Facebook AMP embed
The new Facebook API endpoint doesn't include a trailing slash in the author
URL. This led to a broken URL for the AMP embed of Facebook videos.

For reference the legacy oembed documentation includes it in the example.
https://www.facebook.com/Devex/videos/231586811299265

To see the new result you need to enter an access token and video url here:
https://developers.facebook.com/tools/explorer/?method=GET&path=oembed_video