BAN-2135 Remove unsupported tags from text in apple news format
Sometimes we add scripts in the articles to embed some widgets such as
audio players to have supplementary content in our news pages. These
scripts sometimes have HTML tags that Apple doesn't allow in their
news format. We have to omit them from the text.

This is the list of unsupported tags:
  title
  meta
  script
  noscript
  style
  link
  applet
  object
  iframe
  noframes
  form
  select
  option
  optgroup

https://mydevex.atlassian.net/browse/BAN-2135