BAN-2160 Show image captions on Apple News
The article_json gem is outputting JSON with which Apple News should be
showing image captions in articles.

However these captions were not showing in any form.

This PR makes the necessary adjustments so that captions show in Apple
News articles.

A follow-up PR with deal with the same issue for captions not showing
for videos.

https://mydevex.atlassian.net/browse/BAN-2160