BAN-2160 Show image captions on Apple News for Embed objects
The article_json gem is outputting JSON with which Apple News should be
showing captions in articles. However these captions were not showing in
any form.

This PR follows a similar PR which dealt with captions in images.

This PR makes the necessary adjustments so that captions show in Apple
News articles for `embed` objects.

https://mydevex.atlassian.net/browse/BAN-2160