Fix failing workflow action
Update the source_url for images
Update the article_json to match the updated version number

There has been a change in the urls in the storage from Google. This was
noticeable from our workflow operation on the github page.

All ref_doc changes are global substitutions of these urls.